+++ vhs-frontend/src/App.tsx
@@ -1,6 +1,4 @@
-import React from "react";
 import { Routes, Route } from "react-router-dom";
-
 import Header from "./components/Header";
 import Home from "./pages/Home";
 import Explore from "./pages/Explore";
@@ -8,8 +6,6 @@ import VHSDetails from "./pages/VHSDetails";
 import Edit from "./pages/Edit";
 import CreateNew from "./pages/CreateNew";
 import NotFound from "./pages/NotFound";
-
-import { VHS } from "./types";
 
 import "./App.css";
