+++ vhs-frontend/src/components/DeletePopup.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { MouseEventHandler } from "react";
 import "./popup.css";
