+++ vhs-frontend/src/components/VHSThumbnail.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Link } from "react-router-dom";
 
 type Props = {
