+++ vhs-frontend/src/components/VhsForm.tsx
@@ -25,9 +25,24 @@ export const VhsForm = ({ onSubmit, values }: Props) => {
     },
   });
 
+  const createData = (formData: VhsFormType) => {
+    const data = new FormData();
+    data.append("title", formData.title);
+    data.append("description", formData.description);
+    data.append("genre", formData.genre);
+    data.append("duration", formData.duration.toString());
+    data.append("releasedAt", formData.releasedAt.toString());
+    data.append("rentalPrice", formData.rentalPrice.toString());
+    data.append("rentalDuration", formData.rentalDuration.toString());
+    if (formData.thumbnail) {
+      data.append("thumbnail", formData.thumbnail[0]);
+    }
+    onSubmit(data);
+  };
+
   return (
     <div>
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <form onSubmit={handleSubmit(createData)}>
         <p>Title</p>
         <input {...register("title")} />
         {errors.title && <span>{errors.title.message}</span>}
