+++ vhs-frontend/src/pages/CreateNew.tsx
@@ -1,5 +1,4 @@
 import axios from "axios";
-import { VhsFormType } from "../types";
 import { useNavigate } from "react-router-dom";
 import { VhsForm } from "../components/VhsForm";
 
@@ -19,25 +18,9 @@ export default function CreateNew() {
     }
   };
 
-  const createData = (formData: VhsFormType) => {
-    const data = new FormData();
-    data.append("title", formData.title);
-    data.append("description", formData.description);
-    data.append("genre", formData.genre);
-    data.append("duration", formData.duration.toString());
-    data.append("releasedAt", formData.releasedAt.toString());
-    data.append("rentalPrice", formData.rentalPrice.toString());
-    data.append("rentalDuration", formData.rentalDuration.toString());
-    if (formData.thumbnail) {
-      data.append("thumbnail", formData.thumbnail[0]);
-    }
-    return data;
-  };
-
-  const submitForm = (data: any) => {
+  const submitForm = (data: FormData) => {
     console.log(data);
-    const formData = createData(data);
-    createNewEntry(formData);
+    createNewEntry(data);
     exploreLink("/explore");
   };
