+++ vhs-frontend/src/pages/Edit.tsx
@@ -1,8 +1,7 @@
 import axios from "axios";
-import React, { useState } from "react";
-import { VhsFormType } from "../types";
+import { useState } from "react";
 import { VhsForm } from "../components/VhsForm";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useLocation, useNavigate, redirect } from "react-router-dom";
 import { Popup } from "../components/DeletePopup";
 
 export default function Edit() {
@@ -13,21 +12,6 @@ export default function Edit() {
 
   const config = {
     headers: { "Content-Type": "multipart/form-data" },
-  };
-
-  const createData = (formData: VhsFormType) => {
-    const data = new FormData();
-    data.append("title", formData.title);
-    data.append("description", formData.description);
-    data.append("genre", formData.genre);
-    data.append("duration", formData.duration.toString());
-    data.append("releasedAt", formData.releasedAt.toString());
-    data.append("rentalPrice", formData.rentalPrice.toString());
-    data.append("rentalDuration", formData.rentalDuration.toString());
-    if (formData.thumbnail) {
-      data.append("thumbnail", formData.thumbnail[0]);
-    }
-    return data;
   };
 
   const editEntry = async (data: FormData) => {
@@ -52,8 +36,8 @@ export default function Edit() {
 
   const handleSubmit = (data: any) => {
     console.log(data);
-    const formData = createData(data);
-    editEntry(formData);
+    editEntry(data);
+    redirect(`/${VHS.id}`);
   };
 
   return (
