+++ vhs-frontend/src/pages/Home.tsx
@@ -1,8 +1,3 @@
-import React, { useEffect, useState } from "react";
-import axios from "axios";
-
-type Props = {};
-
-export default function Home({}: Props) {
+export default function Home() {
   return <div>Home</div>;
 }
