+++ vhs-frontend/src/pages/NotFound.tsx
@@ -1,7 +1,3 @@
-import React from "react";
-
-type Props = {};
-
-export default function NotFound({}: Props) {
+export default function NotFound() {
   return <div>NotFound</div>;
 }
